requirements.txt
streamlit
google-generativeai
python-dotenv
werkzeug
flask
flask-cors
PyPDF2
langchain
chromadb
faiss-cpu
langchain_google_genai
regex
langchain-community
Pillow==10.0.0